requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
anyio==3.6.2
apache-airflow==2.5.2
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-mongo==3.0.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
apturl==0.5.2
argcomplete==3.0.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
axios==0.4.0
Babel==2.12.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.1.0
blinker==1.5
boto3==1.26.94
botocore==1.29.94
Brlapi==0.7.0
bs4==0.0.1
cachelib==0.10.2
cattrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
command-not-found==0.3
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.2
cupshelpers==1.0
dataclasses-json==0.5.14
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distro==1.8.0
distro-info===0.23ubuntu1
dnspython==2.3.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.19
duplicity==0.8.12.0
email-validator==1.3.1
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.0
fasteners==0.14.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
future==0.18.2
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httplib2==0.14.0
httpx==0.23.3
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
keyring==18.0.1
langchain==0.0.273
langsmith==0.0.26
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
louis==3.12.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
monotonic==1.5
multidict==6.0.4
mypy-extensions==1.0.0
netifaces==0.10.4
numexpr==2.8.5
numpy==1.24.2
oauthlib==3.1.0
odfpy==1.4.1
olefile==0.46
openai==0.27.9
openpyxl==3.1.2
package-name==0.1
packaging==23.0
pandas==1.5.3
paramiko==2.6.0
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.6.0
Pillow==7.0.0
pkgutil-resolve-name==1.3.10
platformdirs==3.1.1
pluggy==1.0.0
prison==0.2.1
protobuf==3.6.1
psutil==5.9.4
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.36.0
PyJWT==2.6.0
pymacaroons==0.13.0
pymongo==3.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
pyrsistent==0.19.3
python-apt==2.0.1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.21.1
python-http-client==3.3.7
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
pyxdg==0.26
PyYAML==5.4.1
reportlab==3.5.34
requests==2.28.2
requests-toolbelt==0.10.1
requests-unixsocket==0.2.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.2
s3transfer==0.6.0
SecretStorage==2.3.1
sendgrid==6.9.7
setproctitle==1.3.2
simplejson==3.16.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.47
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
ssh-import-id==5.10
starkbank-ecdsa==2.2.0
starlette==0.26.1
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
texttable==1.6.7
tomli==2.0.1
tqdm==4.66.1
twilio==7.16.2
typing-extensions==4.5.0
typing-inspect==0.9.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.1
ufw==0.36
unattended-upgrades==0.1
unicodecsv==0.14.1
urllib3==1.26.15
usb-creator==0.3.7
uvicorn==0.21.1
wadllib==1.3.3
websocket-client==0.59.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
xkit==0.0.0
yarl==1.8.2
zipp==3.15.0